+++ src/hosts/vscode/hostbridge/window/getOpenTabs.test.ts
@@ -62,7 +62,7 @@ describe("Hostbridge - Window - getOpenTabs", () => {
 				return response.paths.length === 2
 			},
 			{
-				timeout: 2000,
+				timeout: 4000,
 				interval: 50,
 			},
 		)
@@ -92,7 +92,7 @@ describe("Hostbridge - Window - getOpenTabs", () => {
 				return response.paths.length === 3
 			},
 			{
-				timeout: 2000,
+				timeout: 4000,
 				interval: 50,
 			},
 		)